Automatically register namespaces for MediaInfo.
This introduces two settings, $wgMediaInfoNamespace = 144 and
$wgMediaInfoTalkNamespace = 145,
that define which namespaces MediaInfo should use. The values set here
are then used to
register MediaInfo with $wgExtraNamespaces, $wgNamespaceContentModels,
and
$wgWBRepoSettings['entityNamespaces'], as appropriate.

To disable this automatic registration, $wgMediaInfoNamespace and
$wgMediaInfoTalkNamespace
can be set to false. In that case, $wgExtraNamespaces,
$wgNamespaceContentModels, and
$wgWBRepoSettings['entityNamespaces'] need to be set explicitly to know
about MediaInfo.

DEPLOY: since the default namespace IDs (144 and 145) may well cause
conflicts, either set
$wgMediaInfoNamespace and $wgMediaInfoTalkNamespace to better values, or
set them to false
and add MediaInfo to $wgExtraNamespaces, $wgNamespaceContentModels, and
$wgWBRepoSettings['entityNamespaces'] explicitly.

Depends-On: I2b7eed07f2b44bdc0aac52d557b42f954318fe3b
Change-Id: Ic0ba92212861e9e4e68a62a9ddd7778fa9e393d1