Allow auto-creation of MediaInfo entities.
A MediaInfo with an ID that corresponds to an existing File page can be
created on the fly by editing (currently, only wbeditentity supports this).

Note: this hooks into a facility created in Wikibase by Ia5f075eb3. However,
this must be merged *before* the change to Wikibase, to avoid CI failures.
Without Ia5f075eb3, this change has no effect.

Bug: T134259
Change-Id: Icaf5f89f1a326e73a45b5227a0d9d49e4446fda4