Remove duplicate dependency on Wikimedia\Assert
The only class that uses this is MediaInfoIdLookup, which is in the
"mediawiki" test set. All these tests depend on MediaWiki anyway (a
dependency that's not declared), which includes the assert class.

Moving this to the require-dev section would be wrong because no class
in the "composer" test set needs this.

Bug: T139343
Change-Id: Id354d87510f42497909f808838eab747e0b95491