Make MediaInfoFieldDefinitions test independent from Wikibase.git code
Tests on this extension are currently failing, because this test was
testing an implementation detail from the two classes it used, but these
classes are in the Wikibase.git code repository.

I needed to remove the specific implementations from the interface to
make this code testable. Note that MediaInfoFieldDefinitions really does
not need to know anything about the implementations. All it needs is the
one method in the interface.

Change-Id: Iea59b8ab0e6324bfa8f62b7dae74c0cc3da1bdf5