Clean up all CSS classnames, prefixing all of ours with "wbmi-"
This aligns us with the general Wikimedia development policy (yay), but
more usefully avoids potential clashes with classnames from other code.
Whilst we're here, I've re-arranged the class definitions a little for
clarity, and dropped some classes that were never/(?)no longer used.

Retain the 'filepage-mediainfo-entitytermsview' class for a week or two,
to allow the caches to cycle over and community gadgets to be updated.

Bug: T213639
Change-Id: I6d04de9b2e42bf9868bb6319cf956ee085f06c43