Don't try to init depicts if it's not enabled
With $wgMediaInfoEnableFilePageDepicts === false, the DOM
nodes for the PHP-rendered depicts data will be absent.
Since they are absent, JS will not manage to attach the
JS depicts widgets. BUT it's trying!
With $wgMediaInfoEnableFilePageDepicts === false, there's
currently a JS notice that JSON.parse (executed a little
lower, on a data attribute of a non-existing node) fails.
We shouldn't even get that far in the first place: as
soon as we detect that nothing's (supposed to be) there,
quit!

Change-Id: Iad074aa57ce1a5fcb82674fd68c6fa80d32a7ba8