Don't update statements that have not changed
This is a pretty massive refactor that introduces Wikibase's
JS data models. The main reason for doing so it that it'll
bring consistency in how the data is used.
The serialized form, as we were using, is not always consistent.
E.g. { entity-type: 'item', id: 'M123' } has the same meaning
as { entity-type: 'item', numeric-id: 123, id: 'M123' }

That made it quite annoying to test if data had changed...

Bug: T216252
Depends-On: I0f2835bd2c225b072e32a1e4fc6cef63f9fb332b
Change-Id: Ia353fa5a091f9f5839cc79049f83b80acc350093