Override oo-ui-inputWidget margin for depicts input element
The .wbmi-depicts-input element (the full-width text input field in the
Depicts panel) was getting a little bit of extra margin on the right side,
keeping it from aligning properly with the "Publish" button above it as the
designs require.

In addition to the `.wbmi-depicts-input` class, this element also receives the
`.oo-ui-inputWidget` class (the source of the excess margin). Overriding the
`margin-right` property that the generic OOUI class provides requires using a
rule with higher specificity; this rule targets any element that is an instance
of both classes. Simply adding `margin-right: 0` to `.wbmi-depicts-input` is
not enough.

Since this specificity override must target a default OOUI style rule by
necessity, I have disabled the selector-class-pattern rule for this line only.

Bug: T216452
Change-Id: Iecfdd9a31ac9343fbb88af45f1f820cf3cebd2c2