Display a "no results found" message in Depicts autocomplete widget
Modifies the EntityLookupElement's getLookupMenuOptionsFromData method to
handle situations where the API request was successful but no results were
found. In this case, the value of the `data` argument will be an empty array.

In this scenario, the method now returns an array containing a disabled
OO.ui.MenuOptionWidget which has a label that reads "No results found", which
is displayed to the user but cannot be selected. Users will see this element
when typing in the Depicts panel autocomplete input if no matching Wikibase
entity is found.

The key for the new label is:
`wikibasemediainfo-filepage-statement-no-results`.

Bug: T217164
Change-Id: Ib909821a9850093155f1c61fef8786025adbd1b4