Move up checks to test if we should construct depicts widgets
Every File: page will load this filepage/init.js file, which will
construct filepage/DepictsPanel.js, which will construct
statements/DepictsWidget.js.

That last one will perform 2 action=wbformatvalue API calls, to
get the label of and link to the "depicts" property.
(These requests are not supposed to go out eventually, they'll
be pre-populated, but because this stuff hasn't yet been
configured, well...)

So right now, we're constructing JS objects that are not yet
needed, will not be attached to DOM, but do fire API calls that
make no sense (because not yet configured)

There already was some code to prevent these widgets to be
attached to DOM and initialized further, but this never should
get constructed in the first place, if these criteria aren't met.
I moved those conditions up a little, so that they now happen
before the entire thing starts to get constructed.

Bug: T217285
Change-Id: I3777783c2f2f9e0ab4f75bbc661d93377da5b226