Basic node setup for front-end unit testing
This patch represents an attempt at getting a front-end testing strategy in
place that relies on node CLI tools instead of Special:JavascriptTest. The
implementation here is inspired by Jdlrobson's proof-of-concept patch in core
(https://gerrit.wikimedia.org/r/c/mediawiki/core/+/487168/) and the test suites
that the Readers Web team has set up in some of their extensions (Popups and
MobileFrontend, for example).

Using a headless node environment for JS unit tests has some advantages over the
traditional browser-based approach:

* Easier to run just the tests we are interested in during the development of
  this extension
* Still compatible with Jenkins & CI pipeline
* Easier to test different use-cases (logged in users vs non-logged-in users,
  for example) in the context of a single method by stubbing/mocking with Sinon
* The ability to `require` a single module makes tests easier to write

This approach does not require any changes in how front-end code is delivered to
users (it does not require the use of Webpack). However, it does require that a
`module.exports` statement be added to any file that will be tested this way, so
that the code can be consumed in a node environment. These exposed modules could
be used in front-end code down the road if desired (thanks to ResourceLoader's
new PackageModules feature) but this is optional.

Description of changes:

* Adds a `test:unit` script to package.json, as well as a series of new dev
  dependencies that are used in the node test suite. This script is included in
  the top-level "test" script, so these tests will run on Jenkins.
* Adds a JS unit test suite in the tests/node-qunit directory. Currently this
  includes:
  - an .eslintrc.json file for the test files (probably needs to be further
    refined)
  - a shim to enable the OOUI library in a headless environment; see the
    comments in that file for more info; ideally this is only something that
    will be needed temporarily
  - a helpers file to assist in mocking various parts of the MW environment
  - adds an example test file that tests the LicenseDialogWidget element
* Adds a `module.exports` line to the LicenseDialogWidget file so that it can be
  consumed as a JS module

To run these tests locally, install the dependencies and then run:

npm run test

Bug: T216229
Change-Id: I540081115818cfb9b9a3a6fd4d297fed74acc4ad