Remove $wgMediaInfoShowQualifiers
This variable was only going to be used as a temporary feature flag
to disable qualifiers for awhile, but the way it's currently
implemented, we don't even need this.

As long as $wgDepictsQualifierProperties is empty (which it is by
default, and none have been configured on beta or prod yet),
qualifiers will not be displayed. $wgMediaInfoShowQualifiers was
only being used to enforce that once more, but, it's pretty useless
(confusing even, because even if $wgMediaInfoShowQualifiers is
enabled, there still won't be any until $wgDepictsQualifierProperties
is configured)

Change-Id: Id27fe42c51b4f71d2818427470ec45cc422db991