Hooks::mediaInfoCaptionsExist: True even if the tag is lower case
CAPTIONS_CUSTOM_TAG is mixed case but it's inserted into the page via
Html::rawElement(), which lower-cases the names of tags "for better
compression". In this test, which we use only to tell the JS code as
to whether there are captions on the page when rendered, we failed to
account for case variance; AFAICT, in all other cases, we were already
using case-invariant tests.

Bug: T220277
Change-Id: If5cb2c50e686362a728424d3f9848a725b555fe4
(cherry picked from commit eb16e4018e7233b26272277e8e529a120a996929)