Handle statements with unsupported data types
We've been operating under the assumption that a statement
is always a wikibase-entityid thing, but that's not necessarily
true. It could be anything else...

The PHP changes here fix the population of the formatValue
cache, which we were doing in a hacky way by creating a fake
snak. That assumed we were dealing with a wiibase-entityid,
which could have another one of those as value. If we're fed
other datatypes, that type of snak is invalid.
Instead, this is going to do the right thing by formatting
just the property value.
(also cleaned up a couple of things in there)

The JS changes fix a minor issue where we're always expecting
the formatted thing to be an item, which has a url. That's
not necessarily true for all data types, and we shouldn't
be trying to construct a link in such case.

Bug: T220613
Bug: T220614
Change-Id: I30bb135d150bf955bbed6bcab13b2fcc1aa6fe2e