Add Selenium tests targeting Beta Commons
Currently it is not feasible to run extension-specific Selenium tests for
MediaInfo in the regular CI process. Instead, tests can be run against Beta
Commons on a regular schedule. These tests need to live in their own
location ("specs_betacommons" instead of "specs") so that they are not picked up
by the Selenium script run by Core (which does happen in the CI pipeline).

This patch:

* Adds a "selenium-test-betacommons" package script to run integration tests
* Adds a series of integration tests at tests/selenium/specs_betacommons
* Adds some documentation on how to run these tests in a Selenium-specific
  README
* Adds the "dotenv" lib as a dev-dependency for configuring selenium, along with
  a pre-configured .env.example file
* Adds the "jpeg-js" lib as a dev-dependency (this is used for random image
  generation in node.js)
* Adds page objects for File page, Login page, Upload page, and Version page
  which are used in tests.

Bug: T219815
Change-Id: I92739d84947ac146ea44176a0785b67bffeeb806
Depends-On: Ib7c3b1b49c280bdc38c40c6aef900a605c1e3cbe