Also render custom statement titles client-size
We're going to have custom titles for more than just depicts.
We can't keep rendering these server-side only, because other
statements will be more flexible: the depicts input box is
always rendered, but other statements won't be until a statement
has been selected.

This means we need to pre-generate & expose even more data in
$jsConfigVars, because those messages need to be made available.
We'll go with the $wgMediaInfoProperties list of properties for
now, but that list might change eventually (I wouldn't be too
surprised if some of this config eventually moves to a wiki page
or so, or maybe something else entirely)

This change also moves the location of the title in the DOM: it
is no longer a direct child of .wbmi-entityview-statementsGroup,
but of one of its children (.wbmi-statements-widget)
There should't be parser cache artifacts: if the cache has the
title in the old spot, JS will wipe out the entire node anyway,
before rerendering the entire widget along with the new title.

Bug: T220214
Change-Id: I1207d346a0607e559e3e47e7a6a7fcafd3ee89ab