Don't show MediaInfo content unless MediaInfo is installed
MediaInfo content on a wiki with MediaInfo enabled works just fine.
However, wikis are configured to share media with Commons when a
file with a given name does not exist locally.
E.g. https://en.wikipedia.org/wiki/File:2019-01-13_Commons_Structured_data_captions.png
pulls in content from https://commons.wikimedia.org/wiki/File:2019-01-13_Commons_Structured_data_captions.png

We may eventually want MediaInfo content to show up on other such
wikis, but we're not in a place where we can easily do so. We'll
need to figure out and work on a couple of things first:
- MediaInfo extension will need to be installed on all such wikis,
  so that the relevant CSS (at least) and JS (preferable) are loaded
- extraction of content from tabs probably needs to change, now
  that content is elsewhere in DOM
- we might even need to change what content goes where (in tabs),
  when part of the content is 'own wiki' and other is from commons
- need to make sure that the JS, if loaded, submits changes to the
  correct wiki (commons, not local)
- I suppose community might need to be consulted about allowing
  edits from other wikis
- etc. etc.

For now, it's best to simply hide the content by default.
MediaInfo will already extract the content, so this is a noop on
MediaInfo-enabled wikis, but at least we won't pollute all other
wikis with content they're not equiped to display (and manage)
properly. Well, the content will still be in DOM, but at least
no-one is going to see it...

There'll only be 1 minor leftover: slots also come with a
h1.mw-slot-header node. MediaInfo's header is empty (and removed
when MI is installed), but it does leave other wikis with an
empty h1.mw-slot-header (which has 1px border-bottom).
I suppose that's something we can live with for now, though...

Bug: T221071
Change-Id: I00a49a1c8f4df3ddb3968979e84f862c8caad1a1
(cherry picked from commit 8027fed2cbbcd3e21d348e6c7bb71cf2521a9ddf)