Don't remove textInput after having submitted it
When submitting multiple captions, this currently removes
the input field for each caption that was submitted by
splicing it out of the array, causing failure after
having submitted half of them (because we've then reached
the end of the array after having spliced half of them
out and reaching half+1 index).

Say we're submitting 3 captions: inputs = [1, 2, 3]
inputs[0] is submitted & the textInput is spliced out.
Remain: inputs = [2, 3] (now with a different index)
inputs[1] is submitted & the textInput is spliced out.
Remain: inputs = [3] (now with a different index)
input[3] is submitted - oh wait a minute, that doesn't
exist!

Until all statements have been submitted, we shouldn't
be altering the array of textInputs (at least not until
it's based on their index, because that'll alter the
indexes)

Bug: T221123
Change-Id: I5b9669eb3c106df5a1ba2cd776dec63d0fab14b5