Update WikibaseMediaInfoHooksTest::testOnWikibaseEntityTypes
This test currently assumes that the 'item' property of the entity type
definitions will not be updated by other extensions in the same hook.
That assumption no longer holds as of the proposed WikibaseCirrusSearch
change in I07f2e2bd34fcb9c637bdb7dffe4e46e74d004e53, which updates the
default UseCirrus config value to true.

Creating and testing for the equality of an 'item' property as done here
is essentially just a test of array_merge, so let's remove it.

Change-Id: I4b5ba300828e755ca8ab9b7f528fa47aab7f0285