tests: Use getNewTempDirectory() without modification
Follows b527b2b2ac9, needed for Iff28c27990a81.

The getNewTempDirectory() method is for creating a new temp directory,
letting MWTestCase track it (for auto-removal) and returning its path.

It is unexpected to modify this path, as it would defeat its purpose.
Let's see if it is still needed. If not, we can discontinue its use
and thus not require core to officially support it.

Depends-On: I7eb4a8c6bf6f6f7103190b9f225579176a7440d6
Change-Id: I4df86f32b7a0a2ca01ff55ab50fecb4d75188c78