Fix vendor versions
Those wikibase modules are not used in production.
The ones used in code running in production is what the Wikibase
extension ships with.
So instead of pinning ourselves to a semver version, we'll want
to make sure tests run again the latest available code, so that
we'll spot breaking changes in time.

Change-Id: I3f52683979533090290fd53a1c8eb8a756c88cb8