Use templates instead of imperative DOM construction when possible
This change is a proof-of-concept and still needs a little bit of work. I'm
submitting it in the hope of having a discussion about the feasibility of doing
this within the MediaInfo extension.

Currently, there are places in the code where we use jQuery methods to
imperatively build out some complex bit of DOM structure. When building out
lots of OOUI widgets, there generally is no other way to do this. But when the
structure consists mostly of static HTML elements (and some conditional classes,
messages, etc.), there is the possibility of moving some of this markup out of
the JS code and into stand-alone template files.

This has several advantages:

  * Mustache templates are already supported
    (https://www.mediawiki.org/wiki/Manual:HTML_templates) and can be used
    without loading any additional libraries.
  * Moving markup out of JS object methods will make classes easier to read
  * Styling and layout changes are much easier to do when all the markpu and CSS
    classes are in one place
  * We could potentially use these templates on the server side as well as the
    client, potentially simplifying our PHP code as well

The disadvantages that I can think of currently are:

  * We need to hook up OOUI widgets manually in some places (some of this still
    needs to be done for the ItemWidget class)
  * We'd need to provide some way to load these templates into a headless
    testing environment, probably by stubbing out calls to `mw.template`.

Change-Id: I074bc2c0cd86023581290305e658912a7af6296b