Make sure statements always get fed the new data
In the case of existing statements, this would not reset their
state when a qualifier was incomplete: the qualifier would then
have no data and it wouldn't be recognized as being different,
even though an (empty) statement input is there.

We should always make sure that, when we reset the data, the
child widgets also get fed that latest data, so they can
reinitialize as needed, even if there's no noticeable change
as far as the parent object is concerned.

This should've been a pretty simple change, where we'd just
feed the new data into the ItemWidget at all times, but due
to the way that one worked, some further refactoring was
involved: ItemWidget would clear out all qualifiers and
reconstruct them (which would resolve in the qualifier widget
causing a delete event, in turn causing the ItemWidget to
emit a change event, which bubbles up and eventually
automatically turns the statement into edit mode, because it
has been 'changed'). To work around that, I had to be a little
more detailed in how qualifiers are handled, only adding/
removing changes and keeping the rest untouched (similar to
how StatementWidget already handles its ItemWidgets)

Bug: T224639
Depends-On: Ib88f2a3d2880c5a6026038dbbd0b52d8275fb2c9
Change-Id: I74da9cf13b5f794ae8d05233827a0c51edd9627e
(cherry picked from commit 769ba097f9a0d722a4bb82d8177dd0c73661e540)