Don't align leading qualifiers line in the middle
It'll now be displayed at 1em instead of in the middle of the
qualifiers element. It used to be in the middle, because the
height of qualifiers in read & edit mode is a little different,
and we'd have a guaranteed decent position here.

However, with qualifiers input fields taking up mulitple lines
on small (mobile) screens (or if qualifier values are long
enough), we don't want that line to show up in the middle, but
around the first line.

Bug: T224624
Depends-On: Ib88f2a3d2880c5a6026038dbbd0b52d8275fb2c9
Change-Id: I497df8872556e4c474670cab3cc7c5212ab4b0c1
(cherry picked from commit 689aa0492ec3c83230cedddb01dd9c66099e4939)