Another approach to detecting the repository of items/properties
So...
- The repository of a items are not included in all of the data
that we have available (e.g. not in JSON blob for existing data)
- And where we have it, it's not always reliable (e.g. repository
for wbsearchentities (directly to wikidata for federation reasons)
is 'local')

The current method of using the formatValue result (which includes
the interwiki prefix in the link's title attribute) works, but
requires that we perform a formatValue API call - something I'd
prefer not to do if we can avoid it (e.g. for items retrieved
from wbsearchentities, where we have the other info)
We assumed that the 'repository' in that API result was correct,
but alas, it is not necessarily (well, 'local' repository does
of course make sense since we're querying wikidata directly...)

So, new approach: fetch the interwiki map and compare the item's
url to it to find the interwiki prefix.
That's just 1 api call, and we can re-use that result for *all*
items (whereas formatValue requires a call *per item*)
I also prefer it over the previous method of using the formatValue
result: that one is less reliable - maybe those things some day
end up being formatted differently...

Bug: T224703
Change-Id: I09d56b2d44e7643effc6fc98b389a1138013dcda