Also support properties with *no* value
This is about the warning we display when users want to edit
a statement with a datatype that is not supported.
We figure that out by checking the data type of the existing
value, but we don't have one for statements added by default
(= depicts) before they get a value.
Since we don't want depicts to show up as unsupported, just
because it doesn't have any data yet, we must also consider
this supported.

Bug: T226096
Change-Id: I111cb91a600750df2a7b500694bd389097bd234e