Fix `data` handling
getData can throw errors when data is invalid/incomplete:
that is already being caught in ItemWidget to ignore invalid
widgets.

Also shouldn't overwrite `this.data` on every change: we
want to keep the original data around for its hash.
this.data is for 'original data passed in' and .getData()
for 'current data as per UI'.

Bug: T226181
Change-Id: Ia1120dcfd369fda86ab2206a1dc643af59d2ce15