Don't change the publish state when marking widget as ready.
`sendData` first makes things read-only (`makeReadOnly`
disables the publish button), but then `setStateReady`
enabled it again. So the next time the widget is set
to edit mode, the button was still enabled.

I suspect earlier versions of the code probably emitted a
change event when it didn't necessarily have to, which would
trigger `onDepictsChange`, which would figure out what state
the button should be in.

Change-Id: I479addc4463bb154f648b81f133c0e5b946a5f63