Fix 'add qualifier' button position when no existing qualifiers
Pre-templates code would render an (empty) container (around the
qualifiers) before the 'add qualifiers' button, which would
push that button down a little.
This is no longer the case - the button should position itself
correctly in that case.

Meanwhile, I'm also removing the no longer needed :empty override.
It was used to hide the container when in read more, when there
were no qualifiers. That is no longer needed, and wouldn't even
work anymore because there will be whitespace inside the container,
and mustache, unlike handlebars, doesn't have syntax for trimming
it.

Change-Id: I01e14a847b813e329765bc7d73bbacd45212b999