Don't let OO.ui.mixin.GroupElement detach nodes from template
This may or may not be the preferred long-term solution...
We may want to look into getting rid of GroupElement complete,
for this precise reasons.
OTOH, it has a few convenient methods for dealing with groups
(move/insert/delete & wired up events, findItemFromData, ...)
that we're currently using...
Let's figure out the best long-term solution later, this need
to get fixed ASAP before this gets deployed.

Bug: T227274
Change-Id: I2083931270a533a5317199d5f757853eb7f0ffb8