Create base before-/afterEach to contruct test environment
Instead of having to wire everything together every individual
test, we can now use these hooks instead.
This will make things safer (don't forget to setup/teardown
anything that might affect the next test), and a lot less
cumbersome to deal with. Instead of having to duplicate a lot
of this boilerplate, it's not mostly consolidated in 1 place.

Change-Id: Iaa309f7f879f016bb1d481d754cdd3f0959f60df