selenium: Remove test for Special:Version
Being listed at Special:Version isn't a feature provided by this
extension. This test isn't triggering any part of the extension,
that a normal page view wouldn't already (e.g. Blankpage).

The integration test to verify that Special:Version works and is
capable of listing extensions belongs in core (and mostly tested
there already).

There is no situation in which this test could fail and lead to
anything other than needing to update the test case due to a
change in core (e.g. the hardcoded DOM elements to select),
or a change in this extension (e.g. label or category).

Previously removed from other extensions as well. I believe
this was originally introduced in early WDIO experiments as an
example test, not as a recommended best practice for every
repo to have a copy of and run.

Change-Id: I9fdaaa8e4fe9f8ebcc8e1c07dd3396871f4441e0