Add LinkNoticeWidget above Statements Panel
Adds a new dismissable UI message within the structured data tab on the file
page. Also adds corresponding i18n messages. A new preference key,
`wbmi-wikidata-link-notice-dismissed` has been added to record whether or not
the user has previously dismissed this message: "0" means they have not, "1"
means they have. These values are kept in localstorage for anon users, and in
user prefs for logged-in users.

To determine whether or not a user should be presented with this widget, first
instantiate it, and then you can call the `isDismissed()` method; the widget
will determine the rest.

A corresponding stylesheet has also been added (there is a small amount of
customization going on).

Finally, tests for the widget's basic functionality have been added. In one
place this led to a minor re-working of how MW user objects are stubbed in
testing (which necessitated making a small change to LicenseDialogWidget's tests
as well).

Bug: T227230
Change-Id: I04ab7dbcfe29a44e76bf92488da3e529a257f066