Don't throw away valid results after abort
Existing code just calls `.abort` on existing promises
when it needs to render something new, so it could mean
we `.abort` a request that has already come back with a
result (e.g. when we want to rerender all the things
for read mode)

We don't need to even delete the value from cache on
`.abort` - aborted API calls will lead to promise
failures and will already be caught. We just need to
make sure not to abort the same thing twice, because
there will only be 1 underlying API call...

Change-Id: I300e43a9a34c23ee7259f612e97395b576cc17da