Don't rerender when property or value changes
When property changes, the value may also need to be
changed (when datatype changes), but that's an OOUI
object, which is not under template control.

Nothing else needs to rerender at this point - that
only needs to happen when we switch back to read mode.

The problem this fixes is the forced rerender when the
value changes during input (e.g. quantity or string input,
notably) would take focus away from those input fields.

Change-Id: I8a0763dc1b3209ca49dcd38831e131cfeaf33c45