Handle additional markup returned by requests to wbformatvalue
Calls to `wbformatvalue` may return additional markup beyond just a link to the
the appropriate property or value. One example is the possible inclusion of a
language fallback indicator <span> element. There may be other possible elements
that could come back as well.

This change updates the render method and template of the QualifierWidget to
handle markup returned from the API. The original markup is preserved, but links
get a target="_blank" added.

Bug: T228693
Change-Id: I6e9877bc7eaf6810b8625de862cdb620adc45c61