Fix formatValue abort handling
There's a race condition when constructing a new NewQualifierWidget
where - because it immediately performs a few renders (construct,
setData & another IIRC) - it aborts previous API calls to format
the value.

That *should* be fine, except that it would re-use the existing
call from cache (the one we just aborted) because the cache was
not cleared soon enough. That's the change in FormatValueElement.

Then, when there are multiple qualifiers of the same property,
it'd still be possible for the last one to abort the formatting
of the first one, resulting in the first one unable to render.

Let's just abort the promises on user input, which is what this
aborting is for anyway (when users input rapidly changing strings),
which should fix the problem of our own code aborting the same
thing over and over - or doing repeat calls for the same thing.

Bug: T228807
Change-Id: Ic3837c92e7ce7f268fbaa5fefd53e87d0647a463