Allow passing in non-default data
StatementWidget.submit() uses state.data to datect changes,
and figure out what to submit to the API.
state.data got populated from setData, which was working just
fine because we didn't have a usecase where we'd want to
programmatically inject new statements (except for the filepage,
but those were what we have in storage, so it it in fact the
correct default state)

Now that we want to provide "copy/sync" functionality in
UploadWizard, we need to make it possible to programatically
set data, and for the submit code to be able to detect that
that data is, in fact, new.

I've changed setData to no longer set the initial data:
it'll now just update the UI etc to reflect the data that
was passed in (like other components), but it will no
longer store that data as "default state".
A (newly renamed) method "resetData) will do the same thing
setData does (update the UI), but it'll also store that
state as default.

So:
- prepopulate with data from storage (or that doesn't need to
  be stored unless changed) -> resetData
- prepopulate with data that needs to be stored -> setData

Meanwhile I also made a few more changes:
- UUID now gets set in ItemWidget instead of in StatementWidget,
  which will make it easier to populate clones
- I've synchronized the timing of some functions that return
  a promise once they're done - they're only really done once
  the other components and functions that they call are also
  complete
- Made stub API calls in testing resolve with an empty object,
  so that they don't stall tests (who'd be waiting for things
  that just don't resolve) now that I've made some promises
  wait for others (that require an API call)

Bug: T231225
Change-Id: I040a5d7f49942cf7fb3838b41f7f2d3d0ecf79c5
(cherry picked from commit 594a63cec04a94e746d89750e26d0eac786a73c3)