Improve DOM merge logic - nodes with different props are not equal
If a node is considered equal, the merge script will just leave it
alone instead of trying to merge the new node into the old DOM.

This works mostly fine, except for nodes that have already have
manual input. isEqualNode seems to only compare attributes (the
values the HTML is constructed with) and not properties (the
actual representation of the current value of the input field)

Which basically means, if we rerender the same node, but with a
different value, it'd currently leave the old one alone since
it consideres them equal (and we want to minimize DOM manipulation)
In this case, that's incorrect, though: if there's a new node with
a new value, we should pick up on that and replace it!

I've also fixed the logic for preserving actual new DOM nodes that
are being passed in (the `originals` array didn't get filled in
time, because the operation that's supposed to fill it
(transformNodes) happened in the same line)

Bug: T231225
Change-Id: I20a20fc99d6a4d39a4e0634e813b75ced5b08231