Skip formatValue calls for input types where we don't need it
For pure input, we don't always need the formatted version &
using the machine representation works just as well.
Strings are the same in both cases anyway.
Quantities are '+50000' in datavalue, and '50,000' formatted:
both of these need some minor manipulation if we're to use
it as input (either trim the + sign, or omit the separator)
For entities, we simply must keep formatting them - the
datavalue only has the id, which is useless to humans without
the formatted item name.

Benefits:
- Fewer API calls for input types where we don't strictly need them
- Easier to test - no need to stub those API calls

Change-Id: I24c2f0dbcc7fbb95e13d7b6731d7e30f045ea93d