Entity lookup: mend old ids when loading
An alternative approach to
Idefcc79f25c4a4e127663e7fa078ef32309b6bf3
but relying on the newly introduced exception.
Given how profound the causing problem is this is merely a band aid and
I am not sure this is really a situation that is recoverable - it is up
to the MediaInfo team to make that call.

Bug: T231276
Depends-On: Ibb73331a619f771cd1ec14ad58c016bf0e587d88
Change-Id: I09c12dcc9793b10f8ac74f6058e248b417d5e0ac