Make MediaInfo implement ClearableEntity
This makes it possible to use the wbeditentity clear parameter on
MediaInfo entities.

The implementation is the same as the one added in change I89a38040a6
(commit 5f5a7a0856), which was later removed in change I2da318df96
(commit 0dcaff1af1) because the need for it was not clear at the time.

Bug: T233522
Change-Id: I0148ae8c76084fdef5f624d7fa6f977b0ef154dc