Perform cross-referencing when finding matching node in next render
To find the matching node in the new render, the current
code will try to find the next best match.

It is possible, however, that *that* next best match is
an even better match for what comes next in the source DOM,
not for the current node (e.g. because it just got deleted,
and the new render is a rerender of all the same things
minus that first one)

Instead of trying the find *the next best match*, we'll now
try to determine all matches ahead of time (before merging
DOMs) and cross-reference matches, so that the order of
appearance (= first wins) no longer matters, but the best
match prevails.

Change-Id: Ice9fa9c978d376956ec0107247464eed443e615d