Bury search code
This search prototype was developed ~2 years ago.
Most of the interesting bits have already been pulled out
of this code and found its way into statements.
It has not seen any change in the last 1.5 years, except for
randomly removing things as we rename/change vars.
This code is not currently functional and would need massive
fixing to bring back.
The prototype has been shot down by the community 1u or so ago.

Let's just put this to rest. There's no point in keeping this dead
code around (it's still in git history) if we have no intention of
ever working on this here, in MediaInfo. It's rotting, it's a
burden, and it adds to the extension & resourceloader payload.

Change-Id: I7d63fe18252ed70e900d80911277e82a776b8817