Fix value-less item rendering
Nothing is currently really broken since actual code always
constructs ItemWidget with a dataValue.
Tests, however, do not. They still accidentally happen to not
fail, but they will throw an exception in the meantime,
because the formatValue call does not receive the argument it
expects.
This just skips the formatValue call in case there is nothing
to format yet.

Change-Id: I6a8866edd11ada6b03364986c912db7532c8c75c