Fix MediaInfoIdLookup
MediaInfoIdLookup was still operating under the assumption that
MediaInfo entities were in their own namespace. This is no longer true.

MediaInfoIds are now composed by prepending 'M' to the file page's
article id - logic that is implemented elsewhere, and instead of
using this existing (but outdated) "lookup", we were composing it
in yet another place - multiple different places, actually.

This fixes MediaInfoIdLookup and consolidates the Title -> entity id
"lookup" in one place, and that class now extends from EntityIdLookup,
an interface that was basically pretty much designed for this (which
should make it possible to use this in other places that already
use this interface)

Change-Id: I12951d59a28d61e4593f00869fe17e4e5df4bd3e